Fix doc issue

+++ 0x00-hello_world/6-size.c
@@ -1,6 +1,5 @@
- /**
+/**
  * main - Print size of variables
- *
  * Description: This function prints size of different variables in the current computer
  * Return: 0
  */
@@ -15,10 +14,10 @@ int main(void)
 	long long int lg2_i;
 	float f;
 
-	printf("Size of a char: %lu byte(s)\n", (unsigned long)sizeof(c));
-	printf("Size of an int: %lu byte(s)\n", (unsigned long)sizeof(i));
-	printf("Size of a long int: %lu byte(s)\n", (unsigned long)sizeof(lg_i));
-	printf("Size of a long long int: %lu byte(s)\n", (unsigned long)sizeof(lg2_i));
-	printf("Size of a float: %lu byte(s)\n", (unsigned long)sizeof(f));
+	printf("Size of a char: %lu byte(s)\n", sizeof(c));
+	printf("Size of an int: %lu byte(s)\n", sizeof(i));
+	printf("Size of a long int: %lu byte(s)\n", sizeof(lg_i));
+	printf("Size of a long long int: %lu byte(s)\n", sizeof(lg2_i));
+	printf("Size of a float: %lu byte(s)\n", sizeof(f));
 	return (0);
 }
